pyproject.toml
[tool.poetry]
name = "core-tutorials"
version = "0.1.0"
description = ""
authors = ["Kevin Liu <kevin.liu4@mail.mcgill.ca>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
django-filter = "^24.2"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
mypy = "^1.11.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[tool.mypy]
ignore_missing_imports = true
exclude = [
    'venv',
    '.venv',
    'env',
    'docs',
    'tests',
]

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_errors = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
